refactor(cli): rename src/harnesses to src/installers
The directory holds CLI installer machinery (ClaudeInstaller,
CodexInstaller, abstract bases, registry). The new @contextbridge/harness
package owns descriptor data. Renaming removes the naming collision and
matches what the files actually are.
