+++ packages/cli/src/installers/ClaudeInstaller.test.ts
