+++ packages/cli/src/installers/ClaudeInstaller.ts
