+++ packages/cli/src/installers/CodexInstaller.test.ts
