+++ packages/cli/src/installers/HarnessInstaller.ts
