+++ packages/cli/src/installers/ScopedHarnessInstaller.ts
