+++ packages/cli/src/installers/detect.test.ts
