+++ packages/cli/src/installers/detect.ts
