+++ packages/cli/src/installers/installers.test.ts
